+++ packages/dsw-tdk/dsw/tdk/consts.py
@@ -11,7 +11,7 @@ APP = 'dsw-tdk'
 PACKAGE_NAME = 'dsw-tdk'
 
 METAMODEL_VERSION_MAJOR = 18
-METAMODEL_VERSION_MINOR = 2
+METAMODEL_VERSION_MINOR = 3
 METAMODEL_VERSION = f'{METAMODEL_VERSION_MAJOR}.{METAMODEL_VERSION_MINOR}'
 
 try:
@@ -29,7 +29,13 @@ REGEX_MIME_TYPE = re.compile(r'^(?![-])(?!.*[-]$)[-\w.]+/[-\w.]+$')
 
 DEFAULT_LIST_FORMAT = '{template.id:<50} {template.name:<30} [{template.uuid}]'
 DEFAULT_ENCODING = 'utf-8'
+DEFAULT_LANGUAGE = 'en'
+DEFAULT_LOCALE_DOMAIN = 'default'
 DEFAULT_README = pathlib.Path('README.md')
+
+JINJA_EXTENSIONS = ('jinja2.ext.do', 'jinja2.ext.loopcontrols')
+JINJA_I18N_TRIMMED = True
+POT_FILE_DEFAULT = 'template.pot'
 
 TEMPLATE_FILE = 'template.json'
 PathspecFactory = pathspec.patterns.GitWildMatchPattern